API/requirements.txt
Django>=4.2
djangorestframework
django-filter
drf-spectacular
django-simple-history
pytest-django

